pyproject.toml
[project]
name = "led"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.0.0",
    "pyhidapi>=0.0.2",
    "pyusb>=1.2.1",
]
